before test in spartan

+++ include/common/plan_common_types.h
@@ -45,7 +45,7 @@ namespace arm_runner {
 
     // The input to command
     struct CommandInput {
-        float control_interval_second;
+        double control_interval_second;
         const RobotArmMeasurement* latest_measurement;
         const RobotCommunication* robot_history;
         const RigidBodyTree<double>* robot_rbt;
